started testing tax vs non tax receipts. Still having problems with postmark
